+++ src/main.ts
@@ -1,3 +1,7 @@
+/**
+ * @packageDocumentation Orchestrates JS and CSS bundle creation in an efficient and configurable manner.
+ */
+
 export { MergeConfigs as MergeRawConfigs } from "./config/merge-configs.js";
 export { ValidateConfig as ValidateRawConfig } from "./config/validate-config.js";
 export {
@@ -7,7 +11,7 @@ export {
     Results,
     ResultsCallback,
 } from "./bundle-orchestrator.js";
-export { BundleStreamFactory } from "./bundle";
+export { BundleStreamFactory, BundleTypes } from "./bundle.js";
 export {
     Bundle,
     Bundles,